refactor, pep-8, todo: setup.py
