+++ unicafe/__init__.py
@@ -1,0 +1,2 @@
+# coding: utf-8
+from unicafe.tools import get_hours, get_food
